Move language server startup to extension root
